requirements.txt
future
emoji
lxml
wget
gtts
beautifulsoup4
requests
python-telegram-bot==13.11
sqlalchemy==1.3.20
psycopg2-binary
feedparser
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
google_images_download
google-api-python-client
googletrans==4.0.0-rc1
google-trans-new==1.1.9
jikanpy
speedtest-cli
coffeehouse
regex
bleach
wikipedia
markdown2
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
pyrogram==1.4.16
PyDictionary
git+https://github.com/Sorrybabe/downloader
youtube_search_python==1.4.9
youtube_search
lyrics_extractor
asyncio
cloudscraper
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
redis
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
tswift
python-arq
envparse
#profanity (Credits to Julia)
better_profanity
textblob 
nudepy
nekos.py
motor
gpytranslate
numpy==1.22.2
fonttools==4.31.2
opencv-python-headless==4.5.5.62
odmantic
tzlocal
pendulum
fuzzysearch
img2pdf
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
git+https://github.com/TotallyNotChase/glitch-this.git
pyYAML>=5.1.2

#ANIBOT
tracemoepy
lottie
natsort
html5lib
AnilistPython